Add navigation.yaml to explain navigation
navigation.yaml should include the hierarchy of the pages, for the
purpose of generating to main navigation, the footer navigation and breadcrumbs.

All pages that will display breadcrumbs in the standard way should be
included in this file. Sections that are to be included in the header
and footer navigations should specify "include_in_menu: True".